+++ src/pkg.hpp
@@ -9,7 +9,7 @@
 struct pkg;
 struct pkg_param;
 
-typedef void (*pkg_dump_pfs_status_t) (std::size_t written, std::size_t total, const char *name, void *ud);
+typedef void (*pkg_dump_pfs_status_t) (std::uint64_t written, std::uint64_t total, const char *name, void *ud);
 
 #define PKG_ENTRY_PARAM_SFO 0x00001000 // param.sfo
 #define PKG_ENTRY_PIC1_PNG  0x00001006 // pic1.png
